✨feat(CadViewer): render cad in svg(WIP) completed functions: * render(lines, arcs, circle) * resize * zoom * center
