+++ src/app/cad-viewer/cad-data/cad-entity/cad-arc.ts
@@ -3,8 +3,10 @@ import {CadLayer} from "../cad-layer";
 import {CadTransformation} from "../cad-transformation";
 import {clampAngle, getVectorFromArray} from "../utils";
 import {CadEntity} from "./cad-entity";
+import {Path} from "@svgdotjs/svg.js";
 
 export class CadArc extends CadEntity {
+	shape?: Path;
 	center: Vector2;
 	radius: number;
 	start_angle: number;
