+++ src/app/components/menu/toolbar/toolbar.component.spec.ts
@@ -7,7 +7,7 @@ import {MatSnackBarModule} from "@angular/material/snack-bar";
 import {provideMockStore} from "@ngrx/store/testing";
 import {initialState} from "@src/app/store/state";
 import {MatMenuModule} from "@angular/material/menu";
-import {CadViewer} from "@src/app/cad-viewer/cad-viewer";
+import {CadViewer} from "@src/app/cad-viewer/cad-viewer-legacy";
 import {CadData} from "@src/app/cad-viewer/cad-data/cad-data";
 import {RouterTestingModule} from "@angular/router/testing";
