+++ src/styles.scss
@@ -5,6 +5,7 @@
 @import "~@syncfusion/ej2-inputs/styles/material.css";
 @import "~jsoneditor/dist/jsoneditor.min.css";
 @import "~highlight.js/styles/nord.css";
+@import "./app/cad-viewer/cad-viewer.scss";
 // @import "./noto-sans-sc.scss";
 
 @font-face {
